Fix README typo.
A note on the phrase here. Technically and legally speaking, you can't prevent someone form sharing the source they make as a derivative work of any item in the project that is under the license. That goes directly against the license itself. The only thing that the license allows you to enforce is that if the copy/derivative work is reshared, it should use the same license or a license that poses the same rights and limits as the publication one. As an example, a derivative work of this repo cannot use the GPL license, since the GPL allows for commercial use.